+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Rational
 
-# Unreleased
+# 0.10.0 (2022-11-28)
 
 - [feat]: `vector` instance can be iterated over it's components;
 - [feat]: providing vector components as arguments to constructor;